Digitalbuzz/Task_1_fb_integration/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.0
psycopg2-binary==2.9.1